Changes and fixes suggested on the PR
